Refactor folder structure (#27)
* Move all source code folders into src

* Update eslintrc to exclude src/pages

* Move images to public folder

* Remove unused image

* Move audio into audio folder

* Move styling related files
