+++ src/pages/_app.js
@@ -7,8 +7,8 @@ import Head from 'next/head'
 import React from 'react'
 
 import Layout from '../components/Layout'
-import createEmotionCache from '../src/createEmotionCache'
-import theme from '../src/theme'
+import theme from '../styles/theme'
+import createEmotionCache from '../utils/createEmotionCache'
 
 // Client-side cache, shared for the whole session of the user in the browser.
 const clientSideEmotionCache = createEmotionCache()
